Fixed Error in put_hex and changed cast type: long double to float: still error

+++ ft_printf.c
@@ -6,7 +6,7 @@
 /*   By: khansman <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2016/05/23 07:54:06 by khansman          #+#    #+#             */
-/*   Updated: 2016/05/26 08:23:40 by oexall           ###   ########.fr       */
+/*   Updated: 2016/05/26 08:41:59 by oexall           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -31,7 +31,7 @@ void				ft_add_param(char type, void *param)
 	else if (type == 'u')
 		ft_put_unum((long unsigned int)param);
 	else if (type == 'f')
-		ft_put_float((long double)param, 6);
+		ft_put_float((float)param, 6);
 }
 
 int					ft_printf(const char *str, ...)
